pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "git-memory"
version = "0.1.0"
description = "AI-powered commit-by-commit memory and structure tracking for Git projects."
readme = "PROJECT.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
   { name = "Dominik Baran" },
]
dependencies = [
   "typer>=0.9.0",
   "rich>=10.0.0",
   "gitpython>=3.1.0",
   "openai>=1.0.0",
   "instructor>=1.0.0",
   "pydantic>=2.0.0",
]

[project.optional-dependencies]
test = [
   "pytest>=7.0.0",
   "pytest-cov>=4.0.0",
   "pytest-mock>=3.10.0",
]

[project.scripts]
git-memory = "git_memory.__main__:app"

[tool.setuptools]
packages = ["git_memory"]
include-package-data = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings",
    "--cov=git_memory",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=80"
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]